# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tindarr API
  version: 1.0.0-draft
  summary: Self-hosted server behind the Tindarr mobile app and web console.
  description: |
    Contract between a Tindarr server and its two clients: the mobile app and the
    web console the server serves under `/`. Written before the implementation; CI
    validates the server's responses against this file and rejects routes it does
    not document.

    Rules (see docs/adr/0006):
    - Additive changes only within v1. Clients ignore unknown fields and treat
      unknown enum values as "other".
    - Errors are `application/problem+json` with a stable `code`.
    - Slow work answers `202` with `retry_after_ms`; the client polls.
    - Every `429` carries `retry_after_ms` and a `Retry-After` header.
    - CORS is disabled: the console is same-origin.

    Every request (docs/auth.md, sections 1 and 2):
    - The `Host` header must be an IP literal, `localhost`, a name in
      `TINDARR_ALLOWED_HOSTS` or the host of `public_url`; anything else gets
      `400` `host_not_allowed` before routing, on every path but `/healthz`.
    - The client IP and scheme come from `X-Forwarded-For` and
      `X-Forwarded-Proto` only when the peer is in `TINDARR_TRUSTED_PROXIES`.
      The server's origin is `<scheme>://<Host>`.
    - Codes, handles and tokens travel in bodies, cookies or headers, never in a
      path or query string.

    Authentication (docs/adr/0009, docs/auth.md section 2), in this order:
...

...
  license:
...
servers:
...
security:
  - bearer: []
  - cookieSession: []
    csrfToken: []
...
paths:
...
  /api/v1/admin/connectors:
    get:
      tags: [admin, console]
      security:
        - cookieSession: []
      summary: All connectors, secrets masked
      operationId: listConnectors
      responses:
        "200":
          description: Connectors.
          content:
            application/json:
              schema:
                type: object
                required: [connectors]
                properties:
                  connectors:
                    type: array
                    items: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
  /api/v1/admin/connectors/{kind}:
    put:
      tags: [admin, console]
      security:
        - cookieSession: []
          csrfToken: []
      summary: Configure a connector
      description: |
        Tested before saving. Omitted secret fields keep their stored value, unless
        the URL changes (`code` = `secret_required`). Fields locked by an
        environment variable cannot be given another value (`code` =
        `setting_locked`).

        `media_server` (docs/auth.md, section 6) needs more than the admin role:
        the caller must be a media server administrator
        (`media_server_admin_required`: a promoted admin cannot repoint the media
        server) with a re-authentication on the **currently configured** media
        server less than 5 minutes old (`reauth_required`, see
        `POST /auth/web/reauth`). No credential is sent to a new URL before it is
        saved. For Plex, `plex_pin_id` is a completed `owner_token` PIN of this
        session, from the account that owns the server; it is consumed only when
        the connector is saved. When the saved server's identity or kind differs
        from the stored one, the same transaction revokes every session, the
        caller's included (the response clears the session cookie), and unlinks
        every user.
      operationId: saveConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorKind"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectorInput" }
      responses:
        "200":
          description: Saved and tested.
          headers:
            Set-Cookie: { $ref: "#/components/headers/ClearSessionCookie" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`admin_required`, `media_server_admin_required`, `reauth_required`, `plex_owner_required` (the PIN's account does not own that Plex server), `csrf_failed` or `remote_access_denied`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: "`not_found`: a connector kind not implemented yet (step 2: every kind but `media_server`)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: "`setting_locked`, `secret_required` or `plex_pin_pending`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "410":
          description: "`pin_expired`: `plex_pin_id` is unknown, expired, used, or not an `owner_token` PIN of this session."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502":
          description: "Connection test failed: `connector_unreachable`, `connector_unauthorized`, `connector_unexpected_response` or `media_server_unsupported`. Nothing is saved."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
    delete:
      tags: [admin, console]
      security:
        - cookieSession: []
          csrfToken: []
      summary: Remove an optional connector (omdb, requests)
      description: |
        Only the optional connectors can be removed. `media_server` is not one of
        them: a Tindarr without a media server signs nobody in, so it is repointed
        with `PUT` instead (`409` `media_server_required`). In step 2 the optional
        connectors do not exist yet and answer `404`.
      operationId: deleteConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorKind"
      responses:
        "204": { description: Removed. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/AdminForbidden" }
        "404":
          description: "`not_found`: a connector kind not implemented yet (step 2: every kind but `media_server`)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: "`setting_locked` (the connector is set by environment variables) or `media_server_required` (the media server connector cannot be removed)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/admin/connectors/{kind}/test:
    post:
      tags: [admin, console]
      security:
        - cookieSession: []
          csrfToken: []
      summary: Test a connector without saving
      description: |
        Returns a coarse result only, never the remote response body. Omitted
        secrets fall back to the stored ones only when the URL is unchanged
        (`code` = `secret_required` otherwise). For a Plex media server,
        `plex_pin_id` is an `owner_token` PIN of this session, which the test does
        not consume. Limited to 10 tests per minute per session.
      operationId: testConnector
      parameters:
        - $ref: "#/components/parameters/ConnectorKind"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectorInput" }
      responses:
        "200":
          description: Test result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectorStatus" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: "`admin_required`, `plex_owner_required` (the PIN's account does not own that Plex server), `csrf_failed` or `remote_access_denied`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: "`not_found`: a connector kind not implemented yet (step 2: every kind but `media_server`)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: "`secret_required` or `plex_pin_pending`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "410":
          description: "`pin_expired`: `plex_pin_id` is unknown, expired, used, or not an `owner_token` PIN of this session."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from an app sign-in, pairing or refresh, valid 15 minutes. Mobile app only.
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_session
      description: |
        Web console session (`kind` = `web`): opaque 256-bit token, stored hashed,
        `HttpOnly`, `Secure`, `SameSite=Strict`, `Path=/`, no `Domain`. Expires
        after 24 h idle or 7 days after sign-in. Named `tindarr_session`, without
        `Secure`, only on requests from a private client IP when
        `TINDARR_ALLOW_HTTP_CONSOLE=true` (docs/auth.md, section 2).
    setupSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_setup
      description: |
        Setup session from `POST /setup/claim` (`kind` = `setup`, no user,
        30 minutes absolute, same attributes as the web session cookie; named
        `tindarr_setup` under the plain-HTTP opt-in). Only one is valid at a time.
        Accepted by the setup operations, `GET /auth/web/session`, the web sign-ins
        while setup is pending, and Plex PINs of purpose `owner_token`. Never
        accepted where a web session is expected, and the reverse.
    preAuth:
      type: apiKey
      in: cookie
      name: __Host-tindarr_preauth
      description: |
        Not a session: a random value that binds the console's Plex PIN and Quick
        Connect sign-in handles to the browser that started them (docs/auth.md,
        section 5). Set by `POST /auth/plex/pins` and `POST /auth/quick-connect`
        when called by the console without `code_challenge`, if absent; 15 minutes;
        cleared by a successful web sign-in. Named `tindarr_preauth` under the
        plain-HTTP opt-in.
    csrfToken:
      type: apiKey
      in: header
...
  parameters:
...
    ConnectorKind:
      name: kind
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ConnectorKind" }
  headers:
...
    ClearSessionCookie:
      description: "Clears `__Host-tindarr_session` (`Max-Age=0`) when the caller's web session was revoked by this request."
      schema: { type: string }
    RetryAfter:
      description: Seconds, rounded up.
      schema: { type: integer, minimum: 0 }
  responses:
...
    BadRequest:
      description: |
        `validation_error` (with `errors`; input values are never echoed),
        `bad_request` (malformed body), or `host_not_allowed` (the `Host` header is
        not an allowed host; any path but `/healthz`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        `unauthorized`: no credential this operation accepts (an `Authorization`
        header on an operation without `bearer`, or a cookie of another kind,
        counts as none), a revoked, idle-expired or past-lifetime session, a
        disabled or unlinked user. `token_expired`: the bearer access token is past
        `exp`; refresh (single-flight) and retry.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    AdminForbidden:
      description: "`admin_required` (the effective role is not `admin`), `csrf_failed` or `remote_access_denied`."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: |
        `rate_limited`, with `retry_after_ms`. Limits are per client IP (or per
        username, session or user where docs/auth.md section 8 says so); global
        thresholds only slow requests down, except the cap on outstanding sign-in
        handles.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: |
            Stable identifier. Known values:
            - request and credentials: validation_error, bad_request,
              host_not_allowed, unauthorized, token_expired, refresh_token_reused,
              csrf_failed, https_required, forbidden, admin_required,
              media_server_admin_required, reauth_required, remote_access_denied,
              not_found, method_not_allowed, rate_limited, internal_error;
            - setup: setup_required, setup_completed, invalid_setup_code,
              setup_session_required, setting_locked;
            - sign-in: invalid_credentials, account_disabled, not_a_server_user,
              password_sign_in_disabled, sign_in_method_unavailable, pin_expired,
              plex_pin_pending, plex_owner_required, plex_tv_unreachable,
              quick_connect_unavailable, quick_connect_expired,
              media_server_unreachable, media_server_changed;
            - pairing: pairing_expired, pairing_rejected,
              pairing_not_awaiting_approval, public_url_not_set,
              public_url_unverified;
            - connectors: connector_unreachable, connector_unauthorized,
              connector_unexpected_response, media_server_unsupported,
              media_server_required, secret_required;
            - swipe and admin: last_admin, streaming_region_not_set,
              metadata_unreachable, llm_not_configured, tmdb_not_configured,
              llm_auth_failed, llm_quota, llm_model_not_found, llm_unreachable,
              llm_invalid_output, daily_limit_reached, requests_not_configured,
              no_backend_user, title_not_offered, request_not_allowed,
              quota_exceeded, request_backend_error.
          examples: [llm_quota]
        retry_after_ms:
          type: integer
          minimum: 0
          description: Set on `rate_limited` and `daily_limit_reached`; the `Retry-After` header carries the same wait in seconds, rounded up.
        reason:
          type: string
          description: |
            Coarse cause for `public_url_unverified`, never a response body:
            `unreachable`, `tls_error`, `redirected`, `unexpected_response` or
            `proof_mismatch`.
          examples: [proof_mismatch]
        errors:
          type: array
          description: Field errors for `validation_error`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    Handle:
      type: string
      description: |
        Plex PIN or Quick Connect handle: 128 random bits, base64url. Never the
        plex.tv PIN id or the Jellyfin secret. Bound to its purpose and initiator,
        in memory only; anything wrong with it gets the same `410` as an unknown
        handle.
      pattern: "^[A-Za-z0-9_-]{22,64}$"
...
    MediaServerKind:
      type: string
      enum: [jellyfin, emby, plex]
    ConnectorKind:
      type: string
      enum: [media_server, requests, tmdb, omdb, llm]
    ConnectorHealth:
      type: string
      description: |
        Coarse result of the last check, never a remote response body.
        `not_configured`: nothing is stored for this connector. `unknown`: it is
        configured but the answer did not test it — `GET /admin/connectors` never
        calls a remote service, so the console asks with
        `POST /admin/connectors/{kind}/test`.
      enum: [ok, unauthorized, unreachable, unexpected_response, unsupported_version, not_configured, unknown]
    ConnectorStatus:
      type: object
      required: [health]
      properties:
        health: { $ref: "#/components/schemas/ConnectorHealth" }
        server_name: { type: [string, "null"] }
        server_version: { type: [string, "null"] }
        checked_at: { type: [string, "null"], format: date-time, description: Null when this answer did not test the connector. }
    SecretState:
      type: object
      description: A stored secret is never returned; only whether it is set.
      required: [set]
      properties:
        set: { type: boolean }
        last4:
          type: [string, "null"]
          description: |
            Last four characters of the stored secret, so an administrator can tell
            two keys apart. Null when showing them would help nobody: a Plex
            connector holds the owner's **account** token, of which there is only
            ever one, so it is masked entirely.
        locked: { type: boolean, description: Set by an environment variable. }
    MediaServerConfigInput:
      type: object
      description: |
        Used by `PUT /setup/media-server` and the `media_server` connector. Jellyfin
        and Emby take `api_key` (an administrator API key), never `plex_pin_id`;
        Plex takes `plex_pin_id`, never `api_key`. An omitted secret keeps the
        stored one only when `url` and `server_type` are unchanged
        (`secret_required` otherwise).
      required: [connector, server_type, url]
      properties:
        connector: { type: string, const: media_server }
        server_type: { $ref: "#/components/schemas/MediaServerKind" }
        url: { type: string, format: uri, maxLength: 512, pattern: "^https?://" }
        api_key: { type: string, minLength: 1, maxLength: 512, description: Jellyfin / Emby admin API key. }
        plex_pin_id:
          $ref: "#/components/schemas/Handle"
          description: Plex only. A completed Plex PIN of purpose `owner_token` from this session; the owner token is taken from it.
        verify_tls:
          type: boolean
          description: Verify the media server's TLS certificate. Omitted means true, and must be omitted when `verify_tls` is locked by the environment.
      if:
        required: [server_type]
        properties:
          server_type: { const: plex }
      then:
        properties:
          api_key: false
      else:
        properties:
          plex_pin_id: false
    LlmProviderKind:
      type: string
      enum: [openai, anthropic, gemini, mistral, openai_compatible, ollama]
    LlmSettingsInput:
      type: object
      required: [connector, provider]
      properties:
        connector: { type: string, const: llm }
        provider: { $ref: "#/components/schemas/LlmProviderKind" }
        api_key: { type: string, maxLength: 512 }
        base_url: { type: string, format: uri, maxLength: 512, description: Required for openai_compatible and ollama. }
        model: { type: string, maxLength: 128 }
        reasoning_effort: { type: [string, "null"], enum: [low, medium, high, null] }
    RequestsInput:
      type: object
      description: Seerr, Jellyseerr or Overseerr.
      required: [connector, url]
      properties:
        connector: { type: string, const: requests }
        url: { type: string, format: uri, maxLength: 512 }
        api_key: { type: string, maxLength: 512 }
        verify_tls: { type: boolean, default: true }
        tv_seasons:
          type: string
          enum: [all, first]
          default: all
          description: Seasons requested for a series.
    ApiKeyInput:
      type: object
      required: [connector]
      properties:
        connector: { type: string, enum: [tmdb, omdb] }
        api_key: { type: string, maxLength: 512 }
    ConnectorInput:
      description: "`connector` must match the `kind` path parameter (`code` = `validation_error` otherwise)."
      oneOf:
        - $ref: "#/components/schemas/MediaServerConfigInput"
        - $ref: "#/components/schemas/LlmSettingsInput"
        - $ref: "#/components/schemas/RequestsInput"
        - $ref: "#/components/schemas/ApiKeyInput"
      discriminator:
        propertyName: connector
        mapping:
          media_server: "#/components/schemas/MediaServerConfigInput"
          llm: "#/components/schemas/LlmSettingsInput"
          requests: "#/components/schemas/RequestsInput"
          tmdb: "#/components/schemas/ApiKeyInput"
          omdb: "#/components/schemas/ApiKeyInput"
    Connector:
      type: object
      required: [kind, configured, status]
      properties:
        kind: { $ref: "#/components/schemas/ConnectorKind" }
        configured: { type: boolean }
        provider: { type: [string, "null"], description: "Media server kind or LLM provider kind." }
        url: { type: [string, "null"] }
        verify_tls: { type: [boolean, "null"] }
        model: { type: [string, "null"] }
        tv_seasons: { type: [string, "null"], enum: [all, first, null], description: Requests connector only. }
        secret: { $ref: "#/components/schemas/SecretState" }
        locked_fields: { type: array, items: { type: string } }
        status: { $ref: "#/components/schemas/ConnectorStatus" }
...